fix a warning.
git-svn-id: https://scm.gforge.inria.fr/authscm/jmcoutey/svn/pm2/trunk@15421 ee107e0b-c2fa-0310-842a-ec2723fb34f4

+++ nmad/drivers/qsnet/src/nm_qsnet.c
@@ -411,7 +411,7 @@ nm_qsnet_accept		(struct nm_cnx_rq *p_crq) {
         {
                 ELAN_EVENT	*ev	= NULL;
                 int tag;
-                u_int size;
+                size_t size;
 
                 ev	= elan_tportRxStart(p, 0, 0, 0, -1, 1, &pkt, sizeof(pkt));
                 elan_tportRxWait(ev, &remote_proc, &tag, &size);
